requirements.txt
fastapi[standard]
beautifulsoup4>=4.12.2
supabase>=2.0.3
lxml>=4.9.3
python-metar